chore: moves eval table definitions and declarative bases into cognee modules + restructures answer executor
