+++ evals/eval_framework/answer_generation/answer_generation_executor.py
@@ -2,38 +2,30 @@ import cognee
 from typing import List, Dict, Callable, Awaitable
 from cognee.api.v1.search import SearchType
 
+question_answering_engine_options: Dict[str, Callable[[str], Awaitable[List[str]]]] = {
+    "cognee_graph_completion": lambda query: cognee.search(
+        query_type=SearchType.GRAPH_COMPLETION, query_text=query
+    ),
+    "cognee_completion": lambda query: cognee.search(
+        query_type=SearchType.COMPLETION, query_text=query
+    ),
+    "cognee_summaries": lambda query: cognee.search(
+        query_type=SearchType.SUMMARIES, query_text=query
+    ),
+    "cognee_insights": lambda query: cognee.search(
+        query_type=SearchType.INSIGHTS, query_text=query
+    ),
+    "cognee_chunks": lambda query: cognee.search(query_type=SearchType.CHUNKS, query_text=query),
+    "cognee_code": lambda query: cognee.search(query_type=SearchType.CODE, query_text=query),
+}
+
 
 class AnswerGeneratorExecutor:
-    # Each option is a function that takes a query (str) and returns an awaitable list of answers.
-    question_answering_engine_options: Dict[str, Callable[[str], Awaitable[List[str]]]] = {
-        "cognee_graph_completion": lambda query: cognee.search(
-            query_type=SearchType.GRAPH_COMPLETION, query_text=query
-        ),
-        "cognee_completion": lambda query: cognee.search(
-            query_type=SearchType.COMPLETION, query_text=query
-        ),
-        "cognee_summaries": lambda query: cognee.search(
-            query_type=SearchType.SUMMARIES, query_text=query
-        ),
-        "cognee_insights": lambda query: cognee.search(
-            query_type=SearchType.INSIGHTS, query_text=query
-        ),
-        "cognee_chunks": lambda query: cognee.search(
-            query_type=SearchType.CHUNKS, query_text=query
-        ),
-        "cognee_code": lambda query: cognee.search(query_type=SearchType.CODE, query_text=query),
-    }
-
     async def question_answering_non_parallel(
-        self, questions: List[Dict[str, str]], qa_engine: str
+        self,
+        questions: List[Dict[str, str]],
+        answer_resolver: Callable[[str], Awaitable[List[str]]],
     ) -> List[Dict[str, str]]:
-        if not questions:
-            raise ValueError("Questions list cannot be empty")
-        if qa_engine not in self.question_answering_engine_options:
-            raise ValueError(f"Unsupported QA engine: {qa_engine}")
-
-        answer_resolver = self.question_answering_engine_options[qa_engine]
-
         answers = []
         for instance in questions:
             query_text = instance["question"]
