+++ evals/eval_framework/corpus_builder/run_corpus_builder.py
@@ -1,11 +1,9 @@
 import logging
 import json
-from uuid import uuid4
-from datetime import datetime, timezone
 from cognee.infrastructure.files.storage import LocalStorage
 from evals.eval_framework.corpus_builder.corpus_builder_executor import CorpusBuilderExecutor
-from evals.eval_framework.corpus_builder.questions_base import QuestionsBase
-from evals.eval_framework.corpus_builder.questions_data import Questions
+from cognee.modules.data.models.questions_base import QuestionsBase
+from cognee.modules.data.models.questions_data import Questions
 from cognee.infrastructure.databases.relational.get_relational_engine import (
     get_relational_engine,
     get_relational_config,
